Prefer 32bit integers for speed

+++ Lib/EEPROM.h
@@ -19,8 +19,8 @@
  * PUBLIC FUNCTIONS
  */
 
-void EEPROM_Write(uint32_t offset, const void * data, uint16_t size);
-void EEPROM_Read(uint32_t offset, void * data, uint16_t size);
+void EEPROM_Write(uint32_t offset, const void * data, uint32_t size);
+void EEPROM_Read(uint32_t offset, void * data, uint32_t size);
 
 /*
  * EXTERN DECLARATIONS
